Moved components into a shared component file
